requirements.txt
openai==1.76.0
streamlit==1.44.1
python-dotenv
pydub
streamlit-audiorecorder==0.0.6
streamlit-javascript==0.1.5